0.2 - WebExtension out of beta
The add-on does not show up in the search results when searching for
"No Native Notifications". Maybe because the non-beta version is a
legacy add-on?

Move the add-on from beta to release to get it to show up.

The add-on has successfully been tested in Firefox 57 on various macOS
versions on Browserstack.